# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Davilla Rondeur e-commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
        - name: featured
          in: query
          schema:
            type: ["boolean", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /orders/by-stripe-session/{stripeSessionId}:
    get:
      operationId: getOrderByStripeSession
      tags: [orders]
      summary: Get order by Stripe Checkout session ID
      parameters:
        - name: stripeSessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    adminBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Product:
      type: object
      required: [id, name, price, categoryId, inStock]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        imageUrl:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        label:
          type: ["string", "null"]
          description: "e.g. Nouveauté, Best-seller"
        inStock:
          type: boolean
        featured:
          type: boolean
        sizes:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
        slug:
          type: string
...
    ShippingAddress:
      type: object
      properties:
        name:
          type: ["string", "null"]
        line1:
          type: ["string", "null"]
        line2:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        postalCode:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
    OrderItem:
      type: object
      required: [id, productId, productName, price, quantity]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productImageUrl:
          type: ["string", "null"]
        price:
          type: number
        quantity:
          type: integer
        size:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
    Order:
      type: object
      required: [id, email, status, items, subtotal, shippingAmount, total, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        status:
          type: string
          enum: [pending, paid, shipped, delivered, cancelled]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: number
        shippingAmount:
          type: number
        total:
          type: number
        shippingAddress:
          oneOf:
            - $ref: "#/components/schemas/ShippingAddress"
            - type: "null"
        stripeSessionId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        paidAt:
          type: ["string", "null"]
          format: date-time
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...