# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Davilla Rondeur e-commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderStatus } from './orderStatus';
import type { ShippingAddress } from './shippingAddress';

export interface Order {
  id: number;
  email: string;
  status: OrderStatus;
  items: OrderItem[];
  subtotal: number;
  shippingAmount: number;
  total: number;
  shippingAddress?: ShippingAddress | null;
  /** @nullable */
  stripeSessionId?: string | null;
  createdAt: Date;
  /** @nullable */
  paidAt?: Date | null;
}