# lib/api-zod/src/generated/types/shippingAddress.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Davilla Rondeur e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface ShippingAddress {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  line1?: string | null;
  /** @nullable */
  line2?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  postalCode?: string | null;
  /** @nullable */
  country?: string | null;
}